refactor: extract the connect function in a separate file
